쿠폰 연결

+++ coursemos_spring_boot/src/main/java/com/example/demo/service/CompanyService.java
@@ -6,7 +6,7 @@ import com.example.demo.domain.Company;
 public interface CompanyService {
 	List<Company> getAllCompany();
 	List<Company> getCompanyByMemberId(Long long1);
-	Company getCompanyByPlaceId(Long placeId);
+	Company getCompanyByPlaceId(Integer placeId);
 	Company getCompanyByCompanyId(Long companyId);
 	void updateCompany(Company company);
 	void insertCompany(Company company);
